requirements.txt
Django==2.0.1
django-tables2==1.17.1
jsonfield==2.0.2
psycopg2==2.7.4
uWSGI==2.0.16
requests==2.18.0
djangorestframework==3.9.0
